+++ src/utils/utils.h
@@ -7,7 +7,7 @@
 
 #include <stddef.h>
 
-void sort_vector(double **a, size_t length);
+void sort_vector(double **a, int length);
 int inside_mesh(double **a, double x, double y, double z, int first, int last);
 
 #endif //MONOALG3D_UTILS_H_H
